Revert settings for testnet. Refactored SearchBar to a separate component.
